dom/base/test/jsmodules/test_multiAsyncImports.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test a loading multiple modules with the async attribute from top level</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var results = [];

  SimpleTest.waitForExplicitFinish();

  function arrayEquals(a, b) {
    if (a.length != b.length) {
      return false;
    }
    for (let i = 0; i < a.length; i++) {
      if (a[i] != b[i]) {
        return false;
      }
      return true;
    }
  }

  function testLoaded() {
    ok(arrayEquals(results.sort(), [1, 2, 3]), 'Check modules imported');
    SimpleTest.finish();
  }
</script>
<script type="module" src="module_simple1.mjs" async></script>
<script type="module" src="module_simple2.mjs" async></script>
<script type="module" src="module_simple3.mjs" async></script>
<body onload='testLoaded()'></body>
